+++ Utilities/Common/CharacterBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
  /*---------------------------------------------------------------------\
  | Copyright © 2008-2011 Allen C. [Alexander Morou] Copeland Jr.        |
  |----------------------------------------------------------------------|
@@ -22,7 +23,7 @@ namespace AllenCopeland.Abstraction.Utilities.Common
 
         }
 
-        public CharacterBase(ulong value)
+        public CharacterBase(BigInteger value)
             : base(controller, value)
         {
         }
@@ -47,12 +48,12 @@ namespace AllenCopeland.Abstraction.Utilities.Common
             }
         }
 
-        protected override NumericBase GetNew(ulong value)
+        protected override NumericBase GetNew(BigInteger value)
         {
             return new CharacterBase(value);
         }
 
-        public static implicit operator CharacterBase(ulong value)
+        public static implicit operator CharacterBase(BigInteger value)
         {
             return new CharacterBase(value);
         }
