+++ Utilities/Common/SexatrigesimalBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
 
 namespace AllenCopeland.Abstraction.Utilities.Common
 {
@@ -14,7 +15,7 @@ namespace AllenCopeland.Abstraction.Utilities.Common
         {
 
         }
-        public SexatrigesimalBase(ulong value)
+        public SexatrigesimalBase(BigInteger value)
             : base(controller, value)
         {
         }
@@ -23,12 +24,12 @@ namespace AllenCopeland.Abstraction.Utilities.Common
         {
         }
 
-        protected override NumericBase GetNew(ulong value)
+        protected override NumericBase GetNew(BigInteger value)
         {
             return new SexatrigesimalBase(value);
         }
 
-        public static implicit operator SexatrigesimalBase(ulong value)
+        public static implicit operator SexatrigesimalBase(BigInteger value)
         {
             return new SexatrigesimalBase(value);
         }
